refactor: streamline codebase by removing utils folder, introducing TelexService interface, and externalizing TELEX_CHANNEL_WEBHOOK configuration i.e taking the webhook url to the properties file
