--- src/main/java/africa/siteanalysisagent/utils/HtmlParserUtil.java
@@ -1,4 +1,0 @@
-package africa.siteanalysisagent.utils;
-
-public class HtmlParserUtil {
-}
