--- src/main/resources/application.properties
@@ -1,5 +1,0 @@
-spring.application.name=site-analysis-agent
-
-spring.main.banner-mode=off
-
-telex.channel.id=01958e7d-78cd-73d4-a9e3-ee05c7a0aab0
