+++ src/main/resources/banner.txt
@@ -1,0 +1,6 @@
+  ____  _ _            _                _           _          _                    _
+ / ___|(_) |_ ___     / \   _ __   __ _| |_   _ ___(_)___     / \   __ _  ___ _ __ | |_
+ \___ \| | __/ _ \   / _ \ | '_ \ / _` | | | | / __| / __|   / _ \ / _` |/ _ \ '_ \| __|
+  ___) | | ||  __/  / ___ \| | | | (_| | | |_| \__ \ \__ \  / ___ \ (_| |  __/ | | | |_
+ |____/|_|\__\___| /_/   \_\_| |_|\__,_|_|\__, |___/_|___/ /_/   \_\__, |\___|_| |_|\__|
+                                          |___/                    |___/
